feat(openai-llm)!: extract OpenAIProvider into @mcp-abap-adt/openai-llm
BaseLLMProvider and LLMProvider types move to core (@mcp-abap-adt/llm-agent)
because they are provider-agnostic abstractions consumed by all LLM-provider
packages and import only types from core.

Server's llm-providers barrel and deepseek.ts now re-import OpenAIProvider
from the new package. Back-compat re-export stays until Task 14.
