+++ packages/llm-agent-server/package.json
@@ -55,6 +55,7 @@
   },
   "dependencies": {
     "@mcp-abap-adt/llm-agent": "*",
+    "@mcp-abap-adt/openai-llm": "*",
     "@modelcontextprotocol/sdk": "^1.28.0",
     "@sap-ai-sdk/orchestration": "^2.9.0",
     "axios": "^1.14.0",
