+++ packages/llm-agent-server/src/agents/prompt-based-agent.ts
@@ -5,8 +5,11 @@
  * The agent returns the raw response to the consumer for any downstream tool handling.
  */
 
-import type { AgentStreamChunk, Message } from '@mcp-abap-adt/llm-agent';
-import type { LLMProvider } from '../llm-providers/base.js';
+import type {
+  AgentStreamChunk,
+  LLMProvider,
+  Message,
+} from '@mcp-abap-adt/llm-agent';
 import {
   type AgentCallOptions,
   BaseAgent,
