+++ packages/llm-agent-server/src/llm-providers/anthropic.ts
@@ -9,8 +9,8 @@ import type {
   LLMResponse,
   Message,
 } from '@mcp-abap-adt/llm-agent';
+import { BaseLLMProvider } from '@mcp-abap-adt/llm-agent';
 import axios, { type AxiosInstance } from 'axios';
-import { BaseLLMProvider } from './base.js';
 
 export interface AnthropicConfig extends LLMProviderConfig {
   model?: string;
