+++ packages/llm-agent/src/index.ts
@@ -1,4 +1,5 @@
 // Re-export everything from core subtrees.
 export * from './interfaces/index.js';
+export { BaseLLMProvider, type LLMProvider } from './llm/base-llm-provider.js';
 export * from './rag/index.js';
 export * from './types.js';
