+++ packages/openai-llm/src/index.ts
@@ -1,0 +1,2 @@
+export type { OpenAIConfig } from './openai-provider.js';
+export { OpenAIProvider } from './openai-provider.js';
