+++ packages/openai-llm/tsconfig.json
@@ -8,5 +8,5 @@
   },
   "include": ["src/**/*"],
   "exclude": ["**/__tests__/**", "**/*.test.ts", "dist"],
-  "references": [{ "path": "../llm-agent" }, { "path": "../openai-llm" }]
+  "references": [{ "path": "../llm-agent" }]
 }
